striatica/pyproject.toml
[project]
name = "striatica"
version = "0.1.0"
description = "A geometric atlas for machine intelligence"
authors = [
    {name = "spacecat",email = "25212708+spacecat@users.noreply.gitlab.com"}
]
requires-python = "^3.12"
dependencies = [
    "numpy (>=2.4.2,<3.0.0)",
    "scipy (>=1.17.1,<2.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "umap-learn (>=0.5.11,<0.6.0)",
    "hdbscan (>=0.8.41,<0.9.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[project.scripts]
striat = "pipeline.cli:main"

[project.optional-dependencies]
ml = [
    "torch (>=2.6.0) ; python_version >= \"3.12\" and python_version < \"3.15\"",
    "sae-lens (>=6.37.0) ; python_version >= \"3.12\" and python_version < \"3.15\"",
    "transformer-lens (>=2.12.0) ; python_version >= \"3.12\" and python_version < \"3.15\"",
    "transformers (>=4.40.0,<5.0.0) ; python_version >= \"3.12\" and python_version < \"3.15\""
]


[[tool.poetry.packages]]
include = "pipeline"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.4,<0.16.0)"
]
